fix a bug with copying text from full screen when using http
